+++ mikeio/__init__.py
@@ -30,8 +30,7 @@ from .dfs0 import Dfs0
 from .dfs1 import Dfs1
 from .dfs2 import Dfs2
 from .dfs3 import Dfs3
-from .dfsu_factory import Dfsu
-from .dfsu import Mesh
+from .dfsu import Mesh, Dfsu
 from .pfs import read_pfs, Pfs, PfsSection, PfsDocument
 from .xyz import read_xyz
 from .dataset import Dataset, DataArray
