+++ mikeio/dfsu/factory.py
@@ -2,8 +2,8 @@ import os
 from mikecore.DfsuFile import DfsuFile, DfsuFileType
 
 from .dfsu import Dfsu2DH
-from .dfsu_layered import Dfsu3D, Dfsu2DV
-from .dfsu_spectral import DfsuSpectral
+from .layered import Dfsu3D, Dfsu2DV
+from .spectral import DfsuSpectral
 
 
 class Dfsu:
