+++ mikeio/dfsu/spectral.py
@@ -7,9 +7,9 @@ from tqdm import trange
 from mikecore.DfsuFile import DfsuFile, DfsuFileType
 
 from .dfsu import _Dfsu
-from .dataset import Dataset, DataArray
-from .dfsutil import _get_item_info, _valid_item_numbers, _valid_timesteps
-from .spectral_utils import plot_2dspectrum, calc_m0_from_spectrum
+from ..dataset import Dataset, DataArray
+from ..dfsutil import _get_item_info, _valid_item_numbers, _valid_timesteps
+from ..spectral import plot_2dspectrum, calc_m0_from_spectrum
 
 
 class DfsuSpectral(_Dfsu):
