+++ mikeio/exceptions.py
