+++ mikeio/spectral.py
